+++ mountpoint-s3/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+### New features
+
+* Add support for running Mountpoint as a specific user with the `--run-as-user` CLI argument. This allows Mountpoint to run as an unprivileged user even when launched by root, such as via fstab or systemd. ([#1439](https://github.com/awslabs/mountpoint-s3/pull/1439))
+
 ## v1.17.0 (May 12, 2025)
 
 ### New features